add a minimal Jupyter notebook replacement
Name is derived from Callisto, a moon of Jupiter

Runs a Python module and watches for changes to that file, including
them as it finds them.  Lines starting with "##" are used to delimit
blocks of code, and only modified blocks are evaluated.

Has hooks to run the Matplotlib event loop in the background so plots
can be interacted with while waiting for more changes.  Need to do
something better here...